refactor: Modal 슬라이드 애니메이션 관련 로직을 커스텀 훅으로 분리
